# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DIMS AI API",
    "version": "1.0.0",
    "description": "API para DIMS AI — plataforma asistida por IA que extrae datos de facturas comerciales y genera la Declaración de Importación de Mercancías Simplificada (DIMS) de la Aduana Nacional de Bolivia. Cubre extracción de facturas, ciclo de vida de la DIMS (crear → editar → generar → validar → exportar/enviar a SUMA), búsqueda de subpartidas del Arancel, productos recurrentes, glosario e integraciones.",
    "contact": {
      "name": "DIMS AI"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.dims.ai/v1",
      "description": "Producción"
    },
    {
      "url": "https://sandbox.api.dims.ai/v1",
      "description": "Sandbox"
    }
  ],
...
  "security": [
    { "bearerAuth": [] },
    { "apiKey": [] }
  ],
  "paths": {
...
    "/dims": {
      "get": {
        "tags": ["DIMS"],
        "summary": "Listar DIMS",
        "description": "Lista paginada de declaraciones, con filtros por estado y búsqueda por proveedor o número.",
        "operationId": "listDims",
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "description": "Filtrar por estado.",
            "schema": { "$ref": "#/components/schemas/DimsEstado" }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Búsqueda por número de DIMS o proveedor.",
            "schema": { "type": "string" }
          },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": {
            "description": "Lista de DIMS.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "pagination"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/DimsResumen" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["DIMS"],
        "summary": "Crear una DIMS",
        "description": "Crea una declaración en estado `borrador`. Puede crearse a partir de una factura ya extraída (facturaId) o desde una subpartida del buscador.",
        "operationId": "createDims",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "facturaId": {
                    "type": "string",
                    "description": "Factura de origen ya extraída.",
                    "example": "fac_8f21a"
                  },
                  "subpartida": {
                    "type": "string",
                    "description": "Subpartida inicial cuando se crea desde el buscador del Arancel.",
                    "example": "8471.30.00.00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "DIMS creada.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Dims" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/dims/{dimsId}": {
      "parameters": [{ "$ref": "#/components/parameters/DimsId" }],
      "get": {
        "tags": ["DIMS"],
        "summary": "Obtener una DIMS",
        "operationId": "getDims",
        "responses": {
          "200": {
            "description": "Detalle de la DIMS.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Dims" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "put": {
        "tags": ["DIMS"],
        "summary": "Actualizar una DIMS",
        "operationId": "updateDims",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DimsUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "DIMS actualizada.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Dims" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "delete": {
        "tags": ["DIMS"],
        "summary": "Eliminar una DIMS",
        "description": "Solo se permite eliminar DIMS en estado `borrador`.",
        "operationId": "deleteDims",
        "responses": {
          "204": { "description": "DIMS eliminada." },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/dims/{dimsId}/generate": {
      "parameters": [{ "$ref": "#/components/parameters/DimsId" }],
      "post": {
        "tags": ["DIMS"],
        "summary": "Generar el formulario y la liquidación",
        "description": "La IA arma el formulario de Aduana y calcula CIF, GA, IVA e ICE según la subpartida de cada ítem.",
        "operationId": "generateDims",
        "responses": {
          "200": {
            "description": "DIMS con liquidación calculada.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Dims" }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/dims/{dimsId}/validate": {
      "parameters": [{ "$ref": "#/components/parameters/DimsId" }],
      "post": {
        "tags": ["DIMS"],
        "summary": "Validar contra SUMA",
        "description": "Comprueba consistencia tributaria, vigencia de subpartidas y restricciones contra el sistema SUMA de la Aduana Nacional.",
        "operationId": "validateDims",
        "responses": {
          "200": {
            "description": "Resultado de la validación.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationResult" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/dims/{dimsId}/export": {
      "parameters": [{ "$ref": "#/components/parameters/DimsId" }],
      "post": {
        "tags": ["DIMS"],
        "summary": "Exportar la DIMS",
        "description": "Genera el archivo de la DIMS en el formato solicitado. Devuelve una URL de descarga temporal.",
        "operationId": "exportDims",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["formato"],
                "properties": {
                  "formato": { "$ref": "#/components/schemas/ExportFormat" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Archivo generado.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExportResult" }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/dims/{dimsId}/email": {
      "parameters": [{ "$ref": "#/components/parameters/DimsId" }],
      "post": {
        "tags": ["DIMS"],
        "summary": "Enviar la DIMS por correo",
        "description": "Envía el archivo de la DIMS (en el formato indicado) como adjunto a un destinatario.",
        "operationId": "emailDims",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["destinatario", "formato"],
                "properties": {
                  "destinatario": {
                    "type": "string",
                    "format": "email",
                    "example": "agente@despachante.bo"
                  },
                  "formato": { "$ref": "#/components/schemas/ExportFormat" }
                }
              }
            }
          }
        },
        "responses": {
          "202": { "description": "Correo encolado para envío." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/dims/{dimsId}/submit": {
      "parameters": [{ "$ref": "#/components/parameters/DimsId" }],
      "post": {
        "tags": ["DIMS"],
        "summary": "Transmitir la DIMS a SUMA",
        "description": "Transmite la declaración directamente al sistema SUMA. La DIMS debe estar validada sin errores. Cambia el estado a `enviada`.",
        "operationId": "submitDims",
        "responses": {
          "200": {
            "description": "DIMS transmitida a SUMA.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Dims" }
              }
            }
          },
          "409": {
            "description": "La DIMS no está en un estado válido para enviar (p. ej. no validada o con errores).",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token de sesión del usuario."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key para integraciones (dims_live_… / dims_test_…)."
      }
    },
    "parameters": {
...
      "DimsId": {
        "name": "dimsId",
        "in": "path",
        "required": true,
        "description": "Número/identificador de la DIMS.",
        "schema": { "type": "string" },
        "example": "DIMS-2026-04823"
      },
      "Page": {
        "name": "page",
        "in": "query",
        "description": "Número de página (1-based).",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "description": "Tamaño de página.",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Solicitud inválida.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "Unauthorized": {
        "description": "No autenticado o credenciales inválidas.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "NotFound": {
        "description": "Recurso no encontrado.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "Conflict": {
        "description": "Conflicto con el estado actual del recurso.",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": { "type": "string", "example": "not_found" },
              "message": { "type": "string", "example": "La DIMS solicitada no existe." },
              "details": {
                "type": "array",
                "items": { "type": "string" }
              }
            }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "required": ["page", "pageSize", "total"],
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "pageSize": { "type": "integer", "example": 20 },
          "total": { "type": "integer", "example": 137 }
        }
      },
      "DimsEstado": {
        "type": "string",
        "enum": ["borrador", "enviada", "aprobada"],
        "description": "Estado del ciclo de vida de la DIMS."
      },
      "DimsResumen": {
        "type": "object",
        "required": ["id", "proveedor", "fecha", "valor", "estado"],
        "properties": {
          "id": { "type": "string", "example": "DIMS-2026-04823" },
          "proveedor": { "type": "string", "example": "Shenzhen Electronics" },
          "fecha": { "type": "string", "format": "date", "example": "2026-05-10" },
          "valor": { "type": "number", "description": "Valor CIF.", "example": 9225.0 },
          "estado": { "$ref": "#/components/schemas/DimsEstado" }
        }
      },
...
      "FacturaItem": {
        "type": "object",
        "required": ["id", "descripcion", "cantidad", "unidad", "precioUnit", "subtotal", "subpartida", "confidence", "aiSuggested"],
        "properties": {
          "id": { "type": "string", "example": "i1" },
          "descripcion": { "type": "string", "example": "Laptop 14\" Intel Core i5, 8GB RAM, 512GB SSD" },
          "cantidad": { "type": "number", "example": 20 },
          "unidad": { "type": "string", "example": "UND" },
          "precioUnit": { "type": "number", "example": 385.0 },
          "subtotal": { "type": "number", "example": 7700.0 },
          "subpartida": {
            "type": ["string", "null"],
            "description": "Subpartida asignada; null si el ítem no está clasificado.",
            "example": "8471.30.00.00"
          },
          "confidence": { "$ref": "#/components/schemas/Confidence" },
          "aiSuggested": {
            "type": "boolean",
            "description": "Indica si la subpartida fue sugerida por la IA.",
            "example": true
          }
        }
      },
...
      "Confidence": {
        "type": "number",
        "minimum": 0,
        "maximum": 100,
        "description": "Nivel de confianza de la IA (0–100). Valores bajos se resaltan para revisión.",
        "example": 94
      },
      "Liquidacion": {
        "type": "object",
        "description": "Cálculo tributario de la DIMS.",
        "properties": {
          "cif": { "type": "number", "description": "Valor CIF en moneda de factura.", "example": 9225.0 },
          "ga": { "type": "number", "description": "Gravamen Arancelario.", "example": 692.5 },
          "iva": { "type": "number", "description": "IVA (tasa efectiva 14.94%).", "example": 1481.42 },
          "ice": { "type": "number", "description": "Impuesto a los Consumos Específicos.", "example": 0 },
          "totalBob": { "type": "number", "description": "Total a pagar en bolivianos.", "example": 15022.16 }
        }
      },
      "Dims": {
        "type": "object",
        "required": ["id", "estado", "proveedor", "fecha"],
        "properties": {
          "id": { "type": "string", "example": "DIMS-2026-04823" },
          "estado": { "$ref": "#/components/schemas/DimsEstado" },
          "facturaId": { "type": "string", "example": "fac_8f21a" },
          "proveedor": { "type": "string", "example": "Shenzhen Electronics Co., Ltd." },
          "fecha": { "type": "string", "format": "date", "example": "2026-05-14" },
          "nit": { "type": "string", "description": "NIT del importador.", "example": "7234182013" },
          "aduanaIngreso": { "type": "string", "example": "030" },
          "regimen": { "type": "string", "example": "IM4" },
          "modalidad": { "type": "string", "example": "SIMPLIFICADA" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FacturaItem" }
          },
          "liquidacion": { "$ref": "#/components/schemas/Liquidacion" },
          "validacion": { "$ref": "#/components/schemas/ValidationResult" },
          "creadaEn": { "type": "string", "format": "date-time" },
          "actualizadaEn": { "type": "string", "format": "date-time" }
        }
      },
      "DimsUpdate": {
        "type": "object",
        "description": "Campos editables de una DIMS en borrador.",
        "properties": {
          "proveedor": { "type": "string" },
          "nit": { "type": "string" },
          "aduanaIngreso": { "type": "string" },
          "regimen": { "type": "string" },
          "modalidad": { "type": "string" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FacturaItem" }
          }
        }
      },
      "ValidationIssue": {
        "type": "object",
        "required": ["nivel", "mensaje"],
        "properties": {
          "nivel": {
            "type": "string",
            "enum": ["error", "advertencia", "info"]
          },
          "campo": { "type": "string", "description": "Campo o ítem relacionado.", "example": "items[2].subpartida" },
          "mensaje": { "type": "string", "example": "La subpartida 8504.40.00.00 requiere certificado adicional." }
        }
      },
      "ValidationResult": {
        "type": "object",
        "required": ["valido", "issues"],
        "properties": {
          "valido": {
            "type": "boolean",
            "description": "true si no hay issues de nivel `error`.",
            "example": false
          },
          "validadaEn": { "type": "string", "format": "date-time" },
          "issues": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationIssue" }
          }
        }
      },
      "ExportFormat": {
        "type": "string",
        "enum": ["xml", "pdf", "json", "print"],
        "description": "xml: XML oficial SUMA · pdf: formato oficial AN · json: estructurado para ERP · print: vista de impresión."
      },
      "ExportResult": {
        "type": "object",
        "required": ["formato", "url"],
        "properties": {
          "formato": { "$ref": "#/components/schemas/ExportFormat" },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "URL temporal de descarga.",
            "example": "https://files.dims.ai/exports/DIMS-2026-04823.xml"
          },
          "nombreArchivo": { "type": "string", "example": "DIMS-2026-04823.xml" },
          "tamanoBytes": { "type": "integer", "example": 24576 },
          "expiraEn": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}